% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSogouScel.R
\name{importSogouScel}
\alias{importSogouScel}
\title{Import a Sogou dictionary.}
\usage{
importSogouScel(strpaths)
}
\arguments{
\item{strpaths}{The path of .scel file.}
}
\value{
A list of: 
 \item{desc}{A data frame of the description.}
 \item{dict}{A data frame of the dictionary.}
}
\description{
Import a scel file of Sogou dictionary.
}
\references{
\url{https://pinyin.sogou.com/dict/}
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
