\name{insertWords}
\alias{insertWords}
\title{Insert new words into dictionary.}
\usage{
  insertWords(strwords,
    analyzer = get("Analyzer", envir = .RwordsegEnv))
}
\arguments{
  \item{strwords}{Vector of words.}

  \item{analyzer}{A JAVA object of analyzer.}
}
\value{
  No results.
}
\description{
  Insert new words into dictionary.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

