% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installDict.R
\name{installDict}
\alias{installDict}
\title{Install a new dictionary.}
\usage{
installDict(dictpath, dictname = "", dictdesc = "")
}
\arguments{
\item{dictpath}{Path of dictionary.}

\item{dictname}{Name of the dictionary. Sogou scel file don't need this input.}

\item{dictdesc}{Description of the dictionary. Default is empty string.}
}
\value{
No results.
}
\description{
Install a new dictionary from a Sogou scel file or text file. Make sure the file encoding is in UTF-8.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
