% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listDict.R
\name{listDict}
\alias{listDict}
\title{List the installed dictionaries.}
\usage{
listDict()
}
\value{
A data frame of: 
 \item{id}{ID of the dictionary.}
 \item{dict}{Name of the dictionary.}
 \item{time}{Created time.}
 \item{size}{Word counts of the dictionary.}
 \item{example}{Example words.}
 \item{desc}{Description of the dictionary.}
}
\description{
List all of the installed user-defined dictionaries.
}
\examples{
listDict()

}
\author{
Jian Li <\email{rweibo@sina.com}>
}
