\name{loadDict}
\alias{loadDict}
\title{Load user defined dictionary.}
\usage{
  loadDict(dictpath = getOption("dic.dir"))
}
\arguments{
  \item{dictpath}{The path of dictionary file or folder .}
}
\value{
  No results.
}
\description{
  Load user defined dictionary.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

