% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSentence.R
\name{parseSentence}
\alias{parseSentence}
\title{Parse a string of text.}
\usage{
parseSentence(text)
}
\arguments{
\item{text}{A vector of strings for parsing.}
}
\value{
A list of: 
 \item{parse}{A data frame of the results of syntactic parsing tree.}
 \item{token}{A data frame of the results of word segmentation.}
 \item{depenencies}{A data frame of the results of dependency parsing.}
}
\description{
Runs the CoreNLP annotators to parse a string of text.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
