\name{PD980105}
\alias{PD980105}
\docType{data}
\title{Corpus of Multi-level Processing for Modern Chinese}
\description{
Corpus from The People's Daily from 1998-01-01 to 1998-01-05.
}
\usage{data(PD980105)}
\format{
  A charactor vector in UTF-8.
}
\references{
\url{http://klcl.pku.edu.cn/gxzy/231686.htm}
}
\keyword{datasets}
