% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setAppDir.R
\name{setAppDir}
\alias{setAppDir}
\title{Set the application path.}
\usage{
setAppDir(appdir)
}
\arguments{
\item{appdir}{The directory path of the application folder. Default is 'tempdir()'.}
}
\value{
No results.
}
\description{
The directory path of the application folder will contain the dictionaries and setting files. You can set a user-defined folder permanently. 
We suggest setting the folder of 'APPDATA' environment variable by running 'setAppDir(\"APPDATA\")'.
}
