% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstallDict.R
\name{uninstallDict}
\alias{uninstallDict}
\title{Uninstall a dictionary.}
\usage{
uninstallDict(dictid)
}
\arguments{
\item{dictid}{The ID of the dictionary, which is shown in the result of \code{\link{listDict}}.}
}
\value{
No results.
}
\description{
Uninstall a user-defined dictionary.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
