\name{deleteWords}
\alias{deleteWords}
\title{Remove words into dictionary.}
\usage{
  deleteWords(strwords,
    analyzer = get("Analyzer", envir = .RwordsegEnv))
}
\arguments{
  \item{strwords}{Vector of words.}

  \item{analyzer}{A JAVA object of analyzer.}
}
\value{
  No results.
}
\description{
  Remove words into dictionary.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

