\name{importSogouScel}
\alias{importSogouScel}
\title{Import Sogou .scel dictionary.}
\usage{
  importSogouScel(strpaths)
}
\arguments{
  \item{strpaths}{Path of dictionary.}
}
\value{
  A string vector of words with attributes "Type" and
  "Description".
}
\description{
  Import Sogou .scel dictionary.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

