% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertWords.R
\name{insertWords}
\alias{insertWords}
\title{Insert new words into analyzer.}
\usage{
insertWords(strwords, analyzer = c("jiebaR", "fmm"))
}
\arguments{
\item{strwords}{Vector of words.}

\item{analyzer}{Which analyzer.}
}
\value{
No results.
}
\description{
When you restart R, all of the wordes will be removed. If you want to keep them please try \code{\link{installDict}}.
}
\author{
Jian Li <\email{rweibo@sina.com}>
}
