\name{installDict}
\alias{installDict}
\title{Install new dictionary.}
\usage{
  installDict(dictpath, transGBK = FALSE)
}
\arguments{
  \item{dictpath}{Path of dictionary.}

  \item{transGBK}{Whether to transform GBK t0 UTF-8.}
}
\value{
  No results.
}
\description{
  Install new dictionary.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

