\name{installDict}
\alias{installDict}
\title{Install new dictionary.}
\usage{
  installDict(dictpath, dictname = "userDefine",
    dicttype = c("text", "scel"))
}
\arguments{
  \item{dictpath}{Path of dictionary.}

  \item{dictname}{English name of the dictionary.}

  \item{dicttype}{Type of the dictionary.}
}
\value{
  No results.
}
\description{
  Install new dictionary.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

