\name{installDict}
\alias{installDict}
\title{Install new dictionary.}
\usage{
  installDict(dictpath, dictname,
    dicttype = c("text", "scel"), load = TRUE)
}
\arguments{
  \item{dictpath}{Path of dictionary.}

  \item{dictname}{English name of the dictionary.}

  \item{dicttype}{Type of the dictionary.}

  \item{load}{Whether to load the dictionary immediately.}
}
\value{
  No results.
}
\description{
  Install new dictionary.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

