\name{uninstallDict}
\alias{uninstallDict}
\title{Uninstall the user defined dictionary.}
\usage{
  uninstallDict(removedict = listDict())
}
\arguments{
  \item{removedict}{Names of the dictionary to be
  uninstalled.}
}
\value{
  No results.
}
\description{
  Uninstall the user defined dictionary.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

