\name{uninstallDict}
\alias{uninstallDict}
\title{Uninstall the user defined dictionary.}
\usage{
  uninstallDict(removedict = listDict()$Name,
    remove = TRUE)
}
\arguments{
  \item{removedict}{Names of the dictionary to be
  uninstalled.}

  \item{remove}{Whether to remove the words immediately.}
}
\value{
  No results.
}
\description{
  Uninstall the user defined dictionary.
}
\author{
  Jian Li <\email{rweibo@sina.com}>
}

